Refactor JSON schema multiple inheritance support.
