+++ codegen/python/fixtures/json_schema/array_type/Email_schema.json
