+++ codegen/python/fixtures/json_schema/array_type/EmailsLong_schema.json
