+++ codegen/python/fixtures/json_schema/array_type/EmailsShort_schema.json
