+++ codegen/python/fixtures/json_schema/simple/User_schema.json
