+++ codegen/python/fixtures/json_schema/struct/Catanimal_schema.json
@@ -1,0 +1,31 @@
+{
+	"$schema": "http://json-schema.org/schema#",
+	"type": "object",
+	"properties": {
+		"cities": {
+			"type": "array",
+			"minItems": 1,
+			"maxItems": 10,
+			"items": {
+				"$ref": "EnumCity_schema.json"
+			}
+		},
+		"colours": {
+			"type": "array",
+			"items": {
+				"type": "string"
+			}
+		},
+		"kind": {
+			"type": "string"
+		},
+		"name": {
+			"type": "string"
+		}
+	},
+	"required": [
+		"kind",
+		"cities",
+		"colours"
+	]
+}
