+++ codegen/python/fixtures/json_schema/struct/UsersPostReqBody_schema.json
@@ -1,0 +1,37 @@
+{
+	"$schema": "http://json-schema.org/schema#",
+	"type": "object",
+	"properties": {
+		"ID": {
+			"type": "string",
+			"minLength": 4,
+			"maxLength": 8
+		},
+		"age": {
+			"type": "integer",
+			"minimum": 16,
+			"maximum": 100,
+			"multipleOf": 4
+		},
+		"grades": {
+			"type": "array",
+			"minItems": 2,
+			"maxItems": 5,
+			"uniqueItems": true,
+			"items": {
+				"type": "integer"
+			}
+		},
+		"item": {
+			"type": "string",
+			"minLength": 2,
+			"pattern": "^[a-zA-Z]+$"
+		}
+	},
+	"required": [
+		"ID",
+		"age",
+		"grades",
+		"item"
+	]
+}
